+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This file is used to list changes made in each version of the openstack-block-st
 * Add default_volume_type config option
 * Update gpfs volume driver path and fix gpfs_images_share_mode initialization
 * Change to cinder v2 service and endpoint
+* Add support for san_password with ibm.storwize_svc.StorwizeSVCDriver
 
 ## 10.0.0
 * Upgrading to Juno